% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-drift.R
\name{condition}
\alias{condition}
\title{Computes the drift model conditional on the allele frequency of one population}
\usage{
condition(model, index, freq)
}
\arguments{
\item{model}{The shared drift model}

\item{index}{The population to condition on}

\item{freq}{The frequency to fix the allele frequency in \code{index} to}
}
\value{
A shared drift model for the remaining populations
}
\description{
Computes the drift model conditional on the allele frequency of one population
}

