% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixturegraphs.R
\name{extract_covariance_matrix}
\alias{extract_covariance_matrix}
\title{Extract a covariance matrix from a fitted admixture graph}
\usage{
extract_covariance_matrix(graph, env, root)
}
\arguments{
\item{graph}{A graph specified using the \code{admixturegraph} package.}

\item{env}{An environment mapping edge names to edge lengths.}

\item{root}{A population we use to root the graph.}
}
\value{
The expected covariance matrix given the graph.
}
\description{
From a graph constructed from the \code{admixturegraph} package, and an environment
containing edge lengths -- as a fit using the \code{admixturegraph} package would provide
-- this function extracts an expected covariance matrix.
}
\details{
The construction of the covariance matrix follows Felsenstein's Inferring Phylogenies (2004)
chapter 23 and needs a leaf to root the graph in. This must be provided by a parameter.
}

