% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-drift.R
\name{joint_drift_model}
\alias{joint_drift_model}
\title{Construct a joint drift model from the mean and covariance matrix
of allele frequencies.}
\usage{
joint_drift_model(mu, Sigma)
}
\arguments{
\item{mu}{The mean allele frequency for each population.}

\item{Sigma}{The covariance matrix of drift.}
}
\value{
A wrapper around the distribution
}
\description{
Construct a joint drift model from the mean and covariance matrix
of allele frequencies.
}

